bot_barber/requirements.txt
python-telegram-bot==20.3
python-telegram-bot[job-queue]
python-dotenv
aiomysql
